Move TargetQueueSettings.java
