+++ worker-message-prioritization-distribution/src/main/java/com/github/workerframework/workermessageprioritization/redistribution/consumption/ConsumptionTargetCalculatorBase.java
@@ -17,7 +17,7 @@ package com.github.workerframework.workermessageprioritization.redistribution.co
 
 import com.github.workerframework.workermessageprioritization.rabbitmq.Queue;
 import com.github.workerframework.workermessageprioritization.targetqueue.TargetQueueSettingsProvider;
-import com.github.workerframework.workermessageprioritization.targetqueue.targetqueuesettings.TargetQueueSettings;
+import com.github.workerframework.workermessageprioritization.targetqueue.TargetQueueSettings;
 
 public abstract class ConsumptionTargetCalculatorBase implements ConsumptionTargetCalculator
 {
