+++ worker-message-prioritization-distribution/src/test/java/com/github/workerframework/workermessageprioritization/redistribution/consumption/ConsumptionTargetCalculatorBaseTest.java
@@ -17,7 +17,7 @@ package com.github.workerframework.workermessageprioritization.redistribution.co
 
 import com.github.workerframework.workermessageprioritization.rabbitmq.Queue;
 import com.github.workerframework.workermessageprioritization.targetqueue.TargetQueueSettingsProvider;
-import com.github.workerframework.workermessageprioritization.targetqueue.targetqueuesettings.TargetQueueSettings;
+import com.github.workerframework.workermessageprioritization.targetqueue.TargetQueueSettings;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
