+++ worker-message-prioritization-target-capacity-calculators/src/main/java/com/github/workerframework/workermessageprioritization/targetqueue/TargetQueueSettings.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.workerframework.workermessageprioritization.targetqueue.targetqueuesettings;
+package com.github.workerframework.workermessageprioritization.targetqueue;
 
 public final class TargetQueueSettings
 {
